digitaltwinfeed/requirements.txt
Click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
WTForms==2.3.1
flask-wtf==0.14.3
flask-flatpages==0.7.2
Markdown==3.2.2
PyYAML==5.3.1
flask-httpauth==4.1.0
flask_restful==0.3.8
pytest==6.0.1
oyaml==1.0
plotly==4.14.1
pandas==1.1.5
numpy==1.19.2
pandas-datareader==0.9.0
yfinance==0.1.59
ta==0.7.0
webcolors==1.11.1
finvizfinance==0.7.2
beautifulsoup4==4.9.3
sec-edgar-downloader==4.0.0
urllib3==1.26.3
xmltodict==0.12.0
sqlalchemy==1.3.23
pyOpenSSL==20.0.1
psycopg2==2.8.6
EIA_python==1.22
yahoo-fin==0.8.8
lxml==4.6.3